refactor(bezier): clean up implicit quadrature module structure
- Rename kernel files with _core suffix to match project convention
- Extract implementation from __init__.py into _implicit_quad.py
- Rename ImplicitPolyQuadrature to ImplicitQuadrature
