+++ src/pantr/bezier/implicit/_bernstein_core.py
