+++ src/pantr/bezier/implicit/_mask_core.py
@@ -30,7 +30,7 @@ import numpy as np
 from numpy import typing as npt
 
 from pantr._numba_compat import nb_jit
-from pantr.bezier.implicit._roots import _restrict_scalar
+from pantr.bezier.implicit._roots_core import _restrict_scalar
 
 M: int = 4
 """Mask grid resolution per axis.
