+++ src/pantr/bezier/implicit/_roots_core.py
